Compile Oriented Programming Complete
