+++ Makefile
@@ -6,38 +6,33 @@
 #    By: sesim <sesim@student.42seoul.kr>           +#+  +:+       +#+         #
 #                                                 +#+#+#+#+#+   +#+            #
 #    Created: 2023/01/27 10:57:54 by sesim             #+#    #+#              #
-#    Updated: 2023/02/03 23:42:14 by sesim            ###   ########.fr        #
+#    Updated: 2023/02/04 21:32:23 by sesim            ###   ########.fr        #
 #                                                                              #
 # **************************************************************************** #
 
 CXX			=	c++
-CXXFLAGS	=	-g3 -Wall -Wextra -Werror -fsanitize=address
-CXXFLAGS_TMP	=	-g3 #-fsanitize=address 
+CXXFLAGS	=	-Wall -Wextra -Werror -fsanitize=address
+CXXFLAGS	=	-g3
+
 RM			=	rm -rf
 
 OBJS_DIR	:=	objs/
 DEPS_DIR	:=	deps/
 
-# SRCS		:=	$(wildcard *.cpp)
-SRCS		:=	main.cpp Server.cpp Receiver.cpp Parser.cpp
+SRCS		:=	$(wildcard *.cpp)
 OBJS		:=	$(addprefix $(OBJS_DIR), $(SRCS:.cpp=.o))
 DEPS		:=	$(addprefix $(DEPS_DIR), $(SRCS:.cpp=.d))
 NAME		:=	ircserv
 
-all :
-	$(CXX) $(CXXFLAGS) $(SRCS) -o $(NAME)
-
-tmp : $(NAME)
+all : $(NAME)
 
 $(NAME) : $(OBJS)
 	$(CXX) $(CXXFLAGS) $^ -o $@
-# $(CXX) $^ -o $@
 
 $(OBJS_DIR)%.o : %.cpp
 	@mkdir -p $(dir $@)
 	@mkdir -p $(DEPS_DIR)
 	$(CXX) $(CXXFLAGS) $< -c -o $@ -MMD
-# $(CXX) $< -c -o $@ -MMD
 	mv $(@:.o=.d) $(DEPS_DIR)
 
 clean:
@@ -46,8 +41,10 @@ clean:
 
 fclean: clean
 	$(RM) $(NAME)
+	$(RM) ircserv.dSYM
 
-re: | fclean all
+re: fclean 
+	@make all
 
 -include $(DEPS)
